Refactor summarizers
Separated out common dependencies between gensim and spacy
